backend/requirements.txt
alembic==1.18.4
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==4.2.1
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.129.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib-metadata==6.11.0
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==4.9.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.9.0
numpy==1.26.4
openai==2.6.0
outcome==1.3.0.post0
packaging>=16.8,<24
passlib[bcrypt]==1.7.4
pandas==2.3.3
pillow==10.4.0
pluggy==1.6.0
protobuf==4.25.8
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
rich==13.9.4
rpds-py==0.27.1
selenium==4.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.46
starlette==0.52.1
streamlit==1.28.1
tenacity==8.5.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.41.0
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.1
webdriver-manager==4.0.1
websockets==16.0
wsproto==1.2.0
zipp==3.23.0
